extract core
