+++ packages/flags/src/lib/core.test.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from 'vitest';
+import { core } from './core';
+
+describe('core', () => {
+  it('should work', () => {
+    expect(typeof core).toBe('function');
+  });
+});
